+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.4.4 (next)
 
+* [#34](https://github.com/dblock/iex-ruby-client/pull/34): Add `IEX::Resource::LargestTrades` - [@gil27](https://github.com/gil27).
 * [#32](https://github.com/dblock/iex-ruby-client/pull/32): Add `IEX::Resource::Sectors` - [@gil27](https://github.com/gil27).
 * Your contribution here.